test(frontend): address orchestration review feedback
