+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions-runtime-policy.test.ts
@@ -1,12 +1,11 @@
-import { beforeEach, describe, expect, test } from "bun:test";
+import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import {
   type AgentSessionRecord,
   type AgentSessionState,
-  appQueryClient,
-  clearAppQueryClient,
   createAdapter,
   createLoadAgentSessions,
   createTaskFixture,
+  createTestQueryClient,
   getSession,
   type LegacyRunSummary,
   OPENCODE_RUNTIME_DESCRIPTOR,
@@ -14,48 +13,26 @@ import {
   type RuntimeInstanceSummary,
   runtimeQueryKeys,
   sessionMessagesToArray,
+  setupDefaultLoadSessionsHost,
 } from "./load-sessions-test-harness";
 
 let legacyHost!: { runsList: (repoPath?: string) => Promise<LegacyRunSummary[]> };
 
 describe("agent-orchestrator load-session runtime policy", () => {
+  let restoreDefaultHost: (() => void) | null = null;
+  let queryClient!: ReturnType<typeof createTestQueryClient>;
+
   beforeEach(async () => {
-    await clearAppQueryClient();
-    const hostModule = await import("../../shared/host");
-    legacyHost = hostModule.host as typeof hostModule.host & {
-      runsList: (repoPath?: string) => Promise<LegacyRunSummary[]>;
-    };
-    hostModule.host.runtimeList = async (repoPath = "/tmp/repo", runtimeKind = "opencode") => [
-      {
-        kind: runtimeKind,
-        runtimeId: "runtime-1",
-        repoPath,
-        taskId: null,
-        role: "workspace",
-        workingDirectory: repoPath,
-        runtimeRoute: {
-          type: "local_http",
-          endpoint: "http://127.0.0.1:4444",
-        },
-        startedAt: "2026-02-22T08:00:00.000Z",
-        descriptor: OPENCODE_RUNTIME_DESCRIPTOR,
-      },
-    ];
-    legacyHost.runsList = async () => [];
-    hostModule.host.runtimeEnsure = async (repoPath) => ({
-      kind: "opencode",
-      runtimeId: "runtime-1",
-      repoPath,
-      taskId: null,
-      role: "workspace",
-      workingDirectory: repoPath,
-      runtimeRoute: {
-        type: "local_http",
-        endpoint: "http://127.0.0.1:4444",
-      },
-      startedAt: "2026-02-22T08:00:00.000Z",
-      descriptor: OPENCODE_RUNTIME_DESCRIPTOR,
-    });
+    queryClient = createTestQueryClient();
+    const hostDefaults = await setupDefaultLoadSessionsHost();
+    legacyHost = hostDefaults.legacyHost;
+    restoreDefaultHost = hostDefaults.restore;
+  });
+
+  afterEach(() => {
+    queryClient.clear();
+    restoreDefaultHost?.();
+    restoreDefaultHost = null;
   });
 
   test("hydrates runtime pending permissions and questions for a requested live session", async () => {
@@ -138,7 +115,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     });
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -251,7 +228,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -328,7 +305,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -423,7 +400,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -546,7 +523,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -572,7 +549,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     });
 
     const queryKey = runtimeQueryKeys.list("opencode", "/tmp/repo");
-    appQueryClient.setQueryData(queryKey, []);
+    queryClient.setQueryData(queryKey, []);
 
     const hostModule = await import("../../shared/host");
     const originalList = hostModule.host.agentSessionsList;
@@ -613,7 +590,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
 
     expect(runtimeEnsureCalls).toEqual([]);
     expect(presenceCalls).toEqual([]);
-    expect(appQueryClient.getQueryState(queryKey)?.isInvalidated).toBe(false);
+    expect(queryClient.getQueryState(queryKey)?.isInvalidated).toBe(false);
     expect(state["external-1"]?.historyHydrationState).toBe("hydrated");
   });
 
@@ -648,7 +625,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
@@ -765,7 +742,7 @@ describe("agent-orchestrator load-session runtime policy", () => {
     };
 
     const loadAgentSessions = createLoadAgentSessions({
-      queryClient: appQueryClient,
+      queryClient,
       activeWorkspace: {
         repoPath: "/tmp/repo",
         workspaceId: "workspace-1",
