+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions-stages-message-merge.test.ts
@@ -1,223 +1,16 @@
 import { describe, expect, test } from "bun:test";
-import type {
-  AgentSessionRecord,
-  RuntimeInstanceSummary,
-  RuntimeKind,
-  TaskCard,
-} from "@openducktor/contracts";
-import { OPENCODE_RUNTIME_DESCRIPTOR } from "@openducktor/contracts";
 import {
-  type AgentEnginePort,
-  type LiveAgentSessionSnapshot,
-  toAgentSessionPresenceSnapshotFromLiveSnapshot,
-} from "@openducktor/core";
-import type { SetStateAction } from "react";
-import { getSessionMessageCount } from "@/state/operations/agent-orchestrator/support/messages";
-import { sessionMessageAt } from "@/test-utils/session-message-test-helpers";
-import type { AgentSessionState } from "@/types/agent-orchestrator";
-import { createAgentSessionPresenceSnapshotFixture } from "../test-utils";
-import {
+  createIntent,
+  createRecord,
+  createSession,
+  createStateHarness,
+  getSessionMessageCount,
   mergeHydratedMessages,
   preparePersistedSessionMergeStage,
-  type SessionLifecycleAdapter,
-  type SessionLoadIntent,
-} from "./load-sessions-stages";
-
-type SessionStateMap = Record<string, AgentSessionState>;
-type AttachSessionInput = Parameters<AgentEnginePort["attachSession"]>[0];
-type ResumeSessionInput = Parameters<AgentEnginePort["resumeSession"]>[0];
-type LiveSnapshotOverrides = Omit<Partial<LiveAgentSessionSnapshot>, "title"> & {
-  title?: string | undefined;
-};
-
-const createSessionSummary = (input: AttachSessionInput | ResumeSessionInput) => ({
-  externalSessionId: input.externalSessionId,
-  role: input.role,
-  startedAt: "2026-03-01T09:00:00.000Z",
-  status: "idle" as const,
-  runtimeKind: input.runtimeKind,
-});
-
-const _createLifecycleAdapter = (
-  overrides: Partial<SessionLifecycleAdapter> = {},
-): SessionLifecycleAdapter => ({
-  hasSession: () => false,
-  listSessionPresence: async () => [],
-  loadSessionHistory: async () => [],
-  attachSession: async (input) => createSessionSummary(input),
-  resumeSession: async (input) => createSessionSummary(input),
-  ...overrides,
-});
-
-const createSession = (overrides: Partial<AgentSessionState> = {}): AgentSessionState => ({
-  externalSessionId: "external-1",
-  taskId: "task-1",
-  repoPath: overrides.repoPath ?? "/tmp/repo",
-  role: "build",
-  status: "idle",
-  startedAt: "2026-03-01T09:00:00.000Z",
-  runtimeKind: "opencode",
-  runtimeId: null,
-  workingDirectory: "/tmp/repo/worktree",
-  messages: [],
-  draftAssistantText: "",
-  draftAssistantMessageId: null,
-  draftReasoningText: "",
-  draftReasoningMessageId: null,
-  contextUsage: null,
-  pendingApprovals: [],
-  pendingQuestions: [],
-  todos: [],
-  modelCatalog: null,
-  selectedModel: null,
-  isLoadingModelCatalog: false,
-  promptOverrides: {},
-  ...overrides,
-});
-
-const createRecord = (overrides: Partial<AgentSessionRecord> = {}): AgentSessionRecord => ({
-  externalSessionId: "external-1",
-  role: "build",
-  startedAt: "2026-03-01T09:00:00.000Z",
-  workingDirectory: "/tmp/repo/worktree",
-  runtimeKind: "opencode",
-  selectedModel: null,
-  ...overrides,
-});
-
-const createIntent = (overrides: Partial<SessionLoadIntent> = {}): SessionLoadIntent => ({
-  repoPath: "/tmp/repo",
-  workspaceId: "workspace-1",
-  taskId: "task-1",
-  mode: "bootstrap",
-  requestedSessionId: null,
-  requestedHistoryKey: null,
-  shouldHydrateRequestedSession: false,
-  shouldReconcileLiveSessions: false,
-  historyPolicy: "none",
-  ...overrides,
-});
-
-const _createSessionPresenceSnapshot = (
-  externalSessionId: string,
-  workingDirectory: string,
-  overrides: LiveSnapshotOverrides = {},
-) =>
-  createAgentSessionPresenceSnapshotFixture({
-    ref: {
-      repoPath: "/tmp/repo",
-      runtimeKind: "opencode",
-      externalSessionId,
-      workingDirectory,
-    },
-    snapshot: {
-      externalSessionId,
-      title: "Builder Session",
-      startedAt: "2026-03-01T09:00:00.000Z",
-      status: { type: "busy" },
-      pendingApprovals: [],
-      pendingQuestions: [],
-      workingDirectory,
-      ...overrides,
-    } as Partial<LiveAgentSessionSnapshot>,
-  });
-
-const _createStalePresence = (externalSessionId: string, workingDirectory: string) =>
-  toAgentSessionPresenceSnapshotFromLiveSnapshot({
-    ref: {
-      repoPath: "/tmp/repo",
-      runtimeKind: "opencode",
-      externalSessionId,
-      workingDirectory,
-    },
-    runtimeId: null,
-    snapshot: null,
-  });
-
-const createStateHarness = (sessions: Record<string, AgentSessionState>) => {
-  let state = sessions;
-  const sessionsRef = { current: state };
-  return {
-    sessionsRef,
-    setSessionsById: (
-      updater:
-        | Record<string, AgentSessionState>
-        | ((current: Record<string, AgentSessionState>) => Record<string, AgentSessionState>),
-    ) => {
-      state = typeof updater === "function" ? updater(state) : updater;
-      sessionsRef.current = state;
-    },
-    updateSession: (
-      externalSessionId: string,
-      updater: (current: AgentSessionState) => AgentSessionState,
-    ) => {
-      const current = state[externalSessionId];
-      if (!current) {
-        return;
-      }
-      state = {
-        ...state,
-        [externalSessionId]: updater(current),
-      };
-      sessionsRef.current = state;
-    },
-    getState: () => state,
-  };
-};
-
-const _createTaskFixture = (): TaskCard => ({
-  id: "task-1",
-  title: "Refactor loader",
-  description: "Split hydration into explicit stages",
-  notes: "",
-  status: "ready_for_dev",
-  priority: 2,
-  issueType: "task",
-  aiReviewEnabled: true,
-  availableActions: [],
-  labels: [],
-  subtaskIds: [],
-  documentSummary: {
-    spec: { has: false },
-    plan: { has: false },
-    qaReport: { has: false, verdict: "not_reviewed" },
-  },
-  agentWorkflows: {
-    spec: { required: false, canSkip: true, available: false, completed: false },
-    planner: { required: false, canSkip: true, available: false, completed: false },
-    builder: { required: true, canSkip: false, available: false, completed: false },
-    qa: { required: false, canSkip: true, available: false, completed: false },
-  },
-  updatedAt: "2026-03-01T09:00:00.000Z",
-  createdAt: "2026-03-01T09:00:00.000Z",
-});
-
-const createRuntime = (
-  workingDirectory: string,
-  runtimeKind: RuntimeKind = "opencode",
-): RuntimeInstanceSummary => ({
-  kind: runtimeKind,
-  runtimeId: "runtime-1",
-  repoPath: "/tmp/repo",
-  taskId: null,
-  role: "workspace",
-  workingDirectory,
-  runtimeRoute: {
-    type: "local_http",
-    endpoint: "http://127.0.0.1:4444",
-  },
-  startedAt: "2026-03-01T09:00:00.000Z",
-  descriptor: OPENCODE_RUNTIME_DESCRIPTOR,
-});
-
-const _createStdioRuntime = (
-  runtimeId: string,
-  workingDirectory: string,
-): RuntimeInstanceSummary => ({
-  ...createRuntime(workingDirectory),
-  runtimeId,
-  runtimeRoute: { type: "stdio", identity: runtimeId },
-});
+  type SessionStateMap,
+  type SetStateAction,
+  sessionMessageAt,
+} from "./load-sessions-stages-test-harness";
 
 describe("load-sessions-stages", () => {
   test("prefers hydrated final assistant messages over stale local streamed rows with the same id", () => {
